requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.1.0
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.7
black==20.8b1
branca==0.4.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.3
channels==2.4.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
decorator==5.1.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.39.1
django-appconf==1.0.4
django-crispy-forms==1.10.0
django-filter==2.4.0
django-googledrive-storage==1.5.2
django-heroku==0.3.1
django-rest-framework==0.1.0
django-tinymce4-lite==1.7.4
django-uni-form==0.9.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
folium==0.12.1.post1
future==0.18.2
geocoder==1.38.1
gevent==20.9.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==0.4.17
gunicorn==20.1.0
httplib2==0.19.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.4
jdcal==1.4.1
Jinja2==3.0.3
jsmin==2.2.2
lazy-object-proxy==1.3.1
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.22.2
oauth2client==4.1.3
oauthlib==3.0.1
openpyxl==3.0.6
packaging==20.9
pathspec==0.8.0
pep8==1.7.1
Pillow==7.2.0
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.2.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.1.0
pytz==2021.1
ratelim==0.1.6
redis==3.5.3
regex==2020.10.28
requests==2.25.1
requests-oauthlib==1.2.0
rsa==4.7
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unipath==1.1
uritemplate==3.0.1
urllib3==1.26.3
websocket==0.2.1
websocket-client==0.57.0
websocket-client2==1.0
websockets==8.1
whitenoise==5.3.0
wrapt==1.11.0
zope.event==4.5.0
zope.interface==5.1.2
